Separate native tests and benchmarks
