+++ nes_py/tests/mappers/test_mapper_000_nrom.py
@@ -1,16 +1,14 @@
-"""Mapper 000 / NROM characterization tests."""
+"""Mapper 000 / NROM public environment tests."""
 
-from nes_py.nes_env import SCREEN_SHAPE_24_BIT
+from nes_py._rom import ROM
 
-from nes_py.tests.mapper_fixtures import chr_bank_marker
-from nes_py.tests.mapper_fixtures import prg_bank_marker
 from nes_py.tests.mappers.common import MapperTestCase
 
 
-class ShouldCharacterizeMapper000NROM(MapperTestCase):
-    """Characterize NROM fixed PRG mapping and CHR ROM behavior."""
+class ShouldLoadMapper000NROM(MapperTestCase):
+    """Exercise NROM through the package API."""
 
-    def test_fixed_32k_prg_mapping_and_chr_rom_reads(self):
+    def test_32k_chr_rom_fixture_constructs_and_steps(self):
         path = self.synthetic_rom(
             'nrom-32k.nes',
             mapper=0,
@@ -18,55 +16,23 @@ class ShouldCharacterizeMapper000NROM(MapperTestCase):
             chr_banks=1,
             mirroring='vertical',
         )
-        env = self.env(path)
 
-        self.assertEqual(prg_bank_marker(0), env._read_prg(0x9000))
-        self.assertEqual(prg_bank_marker(1), env._read_prg(0xd000))
-        self.assertEqual(chr_bank_marker(0), env._read_chr(0x0100))
+        rom = ROM(path)
+        self.assertEqual(0, rom.mapper)
+        self.assertEqual(32, rom.prg_rom_size)
+        self.assertEqual(8, rom.chr_rom_size)
+        self.assertEqual('vertical', rom.mirroring)
+        self.assert_env_smoke(path)
 
-    def test_16k_prg_rom_is_mirrored(self):
+    def test_16k_prg_rom_fixture_constructs_and_steps(self):
         path = self.synthetic_rom(
             'nrom-16k.nes',
             mapper=0,
             prg_banks=1,
             chr_banks=1,
         )
-        env = self.env(path)
 
-        self.assertEqual(prg_bank_marker(0), env._read_prg(0x9000))
-        self.assertEqual(prg_bank_marker(0), env._read_prg(0xd000))
-
-    def test_reset_step_and_render_smoke(self):
-        path = self.synthetic_rom(
-            'nrom-smoke.nes',
-            mapper=0,
-            prg_banks=1,
-            chr_banks=1,
-        )
-        env = self.env(path)
-
-        state = env.reset()
-        self.assertEqual(SCREEN_SHAPE_24_BIT, state.shape)
-        state, reward, done, info = env.step(0)
-        self.assertEqual(SCREEN_SHAPE_24_BIT, state.shape)
-        self.assertIsInstance(reward, float)
-        self.assertIsInstance(done, bool)
-        self.assertIsInstance(info, dict)
-        self.assertEqual(SCREEN_SHAPE_24_BIT, env.render('rgb_array').shape)
-
-    def test_prg_ram_is_mapper_state_across_backup_restore(self):
-        path = self.synthetic_rom(
-            'nrom-prg-ram.nes',
-            mapper=0,
-            prg_banks=1,
-            chr_banks=1,
-        )
-        env = self.env(path)
-
-        env._write_prg(0x6000, 0x2a)
-        env._backup()
-        env._write_prg(0x6000, 0x7f)
-
-        self.assertEqual(0x7f, env._read_prg(0x6000))
-        env._restore()
-        self.assertEqual(0x2a, env._read_prg(0x6000))
+        rom = ROM(path)
+        self.assertEqual(0, rom.mapper)
+        self.assertEqual(16, rom.prg_rom_size)
+        self.assert_env_smoke(path)
